requirements.txt
google-adk>=1.28.0
google-generativeai>=0.8.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
streamlit>=1.31.0
authlib>=1.3.0
httpx>=0.26.0
PyJWT>=2.8.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
itsdangerous>=2.1.2
requests>=2.31.0